Last Sync: 2026-06-25 20:18 (Mobile)

+++ V2rayNG/app/src/debug/java/com/v2ray/ang/ui/debug/DebugBadgeController.kt
@@ -27,7 +27,7 @@ object DebugBadgeController {
             setText(BADGE_TEXT)
             setTextColor(colorOnPrimary)
             setSlantedBackgroundColor(colorPrimary)
-            setSlantedLength(28 * density)
+            setSlantedLength((28 * density).toInt()) 
             setMode(SlantedTextView.MODE_RIGHT)
             setTextSize(11)
         }
